Migrated to Linux/cygwin 1.Includes and Definitions   DWORD, LONG, BYTE; HANDLE, PVOID, ULONG_PTR   __fastcall   some secure (_s) functions   socket address structure difference 2.atomic memory operations/intrinsics   note that 'long' integer width is not portable   padding by alignment extension is not portable 3.IPC   message queues   named shared memory vs. anonymous filemapping   exploit high-precision event timer to poll instead of named event 4.Timers   Create a Per-Process Timer, P1003.1b-1993, p. 264 5.RFC2292 compatibility 6.Misc Functions   Sleep   strncasecmp   GetTickCount64   BitScanForward, _bittestandset 7.Miscellaneous   Goto jump over pointer initialization   Variadic Template   fd_set: poll vs select   light weight lock 8.Exploit CMake build management system

+++ src/FSP_FlowTest/Test_TryWideChar.cpp
@@ -55,12 +55,12 @@ void TryWideChar()
 	FindClose(hFind);
 
 	//parsing the buffer
-	int len = min(p - p0, MAX_BLOCK_SIZE);
+	int len = (int)min(p - p0, MAX_BLOCK_SIZE);
 	do
 	{
 		ParseBlock(p0, len);
 		p0 += len;
-		len = min(p - p0, MAX_BLOCK_SIZE);
+		len = (int)min(p - p0, MAX_BLOCK_SIZE);
 	} while (len > 0);
 }
